requirements.txt
pandas
matplotlib
openpyxl
reportlab
thinker
Pillow
folium
tk
customtkinter